# openapi.yaml
openapi: "3.0.0"

info:
  title: Hospital Management Service
  version: "1.0.0"
servers:
  - url: https://hms-3m25jovmiq-uc.a.run.app/

paths:
  /users/patient:
    post:
      tags:
        - Users
      description: To create User account
      operationId: CreateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/schemas/CreateUser"
      responses:
        "201":
          description: Create User
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "404":
          description: Failed to create user
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              

  /users/patient_signin:
    post:
      tags:
        - User
      description: For SignIn user to the DB
      operationId: SignIn
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: SignIn successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                  id:
                    type: string

                  token: 
                    type: string
        "404":
          description: SignIn bad request   
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string
        "500":
          description: Sign Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string

  /user/patient_profile:
    get:
      tags:
        - User
      description: Get the patient profile data
      operationId: GetProfile
      parameters:
        - in: query
          name: id
          schema:
            type: string
      responses:
        "200":
          description: Patient profile data retrival data
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  email:
                    type: string
                  gender:
                    type: string
                  mobile:
                    type: string
                  address:
                    type: string
                  age:
                    type: number
                  medical_history:
                    type: string
        "404":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string
          
  

  /user/update_patient:
      put:
        tags:
          - User
        description: Update User data
        operationId: Updatepatient
        parameters:
          - in: query
            name: id
            schema:
              type: string
        responses:
          "200":
            description: Successful response
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
          "404":
            description: Bad request
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
  
  /doctor:
      post:
        tags:
          - doctor
        description: Create doctor
        operationId: CreateDoctor
        requestBody:
          content:
            application/json:
              schema:
                $ref: "#/schemas/CreateDoctor"
        responses:
          "201":  
            description: Doctor Created successfully
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message: 
                      type: string
          "404":
            description: Bad request
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message: 
                      type: string
  /doctor/signin:
      post:
        tags:
          - doctor
        description: For patient signin
        operationId: DoctorSign
        requestBody:
          content:
            application/json:
              schema:
                type: object
                properties:
                  email: 
                    type: string
                  password: 
                    type: string
        responses:
          "200":
            description: "Successfully signin"
            content:
              application/json:
                schema:
                  type: object
                  properties: 
                    id:
                      type: string
                    email:
                      type: string
                    name: 
                      type: string
                    mobile:   
                      type: string
                    specialization:
                      type: string
          "404":
            description: Bad request
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
          "500":
            description: Internal server error
            content:
              application/json:
                schema:
                  type: object
                  properties:
                  message:
                    type: string      

  /users/getdata:
    get:
      tags:
        - doctor
      description: To get doctor
      operationId: GetDoctor
      parameters:
        - in: query
          name: id
          schema:
            type: string
      responses:
        "200":
          description: Successfully data retrieved
          content:
            application/json:
              schema:
                $ref: "#/schemas/GetDoctorData"
        "404":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string


  /appointment:
      post:
        tags:
          - appointments
        description: Schedule appointment
        operationId: CreateAppointment
        requestBody:
          content:
            application/json:
              schema:
                $ref: '#/schemas/CreateAppointment'
        responses:
        
          "201":  
            description: Successfully Appointment Created
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string

          "404":
            description: Bad Request
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string

          "500":
            description: Internal server error
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
      
schemas:
  CreateUser:
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      password:
        type: string
      gender:
        type: string
      mobile:
        type: string
      address:
        type: string
      age:
        type: number
      medical_history:
        type: string

  CreateDoctor:
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      password:
        type: string
      address:
        type: string
      mobile:
        type: string
      specialization:
        type: string

  CreateAppointment:
    type: object
    properties:
      patient_id:
        type: string
      doctor_id:
        type: string
      appointment_date:
        type: string
      appointment_time:
        type: string

  GetDoctorData:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      email:
        type: string
      mobile:
        type: string
      address:
        type: string

      specialization:
        type: string